Moved GenericField to cloudify-ui-components
